+++ lib/libc/hash/helper.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: helper.c,v 1.15 2015/11/01 03:45:29 guenther Exp $ */
+/*	$OpenBSD: helper.c,v 1.16 2016/09/21 04:38:57 guenther Exp $ */
 
 /*
  * Copyright (c) 2000 Poul-Henning Kamp <phk@FreeBSD.org>
@@ -82,7 +82,7 @@ HASHFileChunk(const char *filename, char *buf, off_t off, off_t len)
 	}
 
 	while ((nr = read(fd, buffer, MINIMUM(sizeof(buffer), len))) > 0) {
-		HASHUpdate(&ctx, buffer, (size_t)nr);
+		HASHUpdate(&ctx, buffer, nr);
 		if (len > 0 && (len -= nr) == 0)
 			break;
 	}
@@ -97,7 +97,7 @@ DEF_WEAK(HASHFileChunk);
 char *
 HASHFile(const char *filename, char *buf)
 {
-	return (HASHFileChunk(filename, buf, (off_t)0, (off_t)0));
+	return (HASHFileChunk(filename, buf, 0, 0));
 }
 DEF_WEAK(HASHFile);
